LayoutTests/accessibility/mac/search-predicate-for-adhoc-radio-groups-expected.txt

This test ensures WebKit can search for ad-hoc radiogroups. An ad-hoc radiogroup is a set of radiobuttons connected by `name` attribute alone, without an appropriate `role="radiogroup"` container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resultElement.role is 'AXRole: AXRadioButton'
PASS resultElement.domIdentifier is 'circle'
PASS resultElement.role is 'AXRole: AXRadioButton'
PASS resultElement.domIdentifier is 'red'
PASS resultElement.role is 'AXRole: AXRadioButton'
PASS resultElement.domIdentifier is 'small'
PASS resultElement.role is 'AXRole: AXRadioButton'
PASS resultElement.domIdentifier is 'green'
PASS resultElement.role is 'AXRole: AXRadioButton'
PASS resultElement.domIdentifier is 'triangle'
PASS resultElement.role is 'AXRole: AXRadioButton'
PASS resultElement.domIdentifier is 'red'
PASS successfullyParsed is true

TEST COMPLETE
Shape

 Circle
 Square
 Triangle
Color

 Red
 Blue
 Green
Size

 Small
Unrelated content separating ad-hoc radio group items

 Medium
Unrelated content separating ad-hoc radio group items

 Large
